fix: ensure Ruby VM scheduling thread stopped
There was some race condition in between
forking Ruby VM with Process.fork+stop_timer_thread
and moving a process into new User namespace.
The attach() op was failing because after fork
there was some thread out of the blue that I was
not aware about. It reproduced. Not often.

The root cause appeared to be a race in between
the start of the Ruby scheduling thread and
its stop. Start is not a sync op and rarely
it overtakes the stop signal.

The solution is simple - after we issued a stop
signal the program must wait till the last thread
disappear.

There is also a mitigation of this
failure on multiple levels - retries and restarts
are possible to do in both setns() calls and on
the top level lxc-attach() call.